Desktop/upwork_dashboard/backend/requirements.txt
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
google-generativeai==0.3.2
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
